# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenNotebook",
    "description": "Collections of sources, and the decks, audio overviews, mind maps and notes made from them. Everything the web app does goes through this API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/shares": {
      "get": {
        "tags": [
          "shares"
        ],
        "summary": "List Shares",
        "description": "Discover: what everyone on the studio shared, a page at a time. When\nthere are more after this page, the answer's `X-Next-Offset` header is\nthe `offset` to ask for next.",
        "operationId": "list_shares",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "description": "Case-insensitive; matched against the title, the note and the cover's topic and terms",
              "default": "",
              "title": "Query"
            },
            "description": "Case-insensitive; matched against the title, the note and the cover's topic and terms"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "newest",
                "reused"
              ],
              "type": "string",
              "description": "`newest` (the default) or `reused`: most reused first",
              "default": "newest",
              "title": "Sort"
            },
            "description": "`newest` (the default) or `reused`: most reused first"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "The most shares in one answer",
              "default": 100,
              "title": "Limit"
            },
            "description": "The most shares in one answer"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "description": "Where this page starts: the `X-Next-Offset` header of the page before",
              "default": 0,
              "title": "Offset"
            },
            "description": "Where this page starts: the `X-Next-Offset` header of the page before"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShareCard"
                  },
                  "title": "Response List Shares"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/items": {
      "get": {
        "tags": [
          "shares"
        ],
        "summary": "List Shared Items",
        "description": "Discover's items: each ready output a share includes, on its own, from\nthe shares of everyone on the studio. Only what a share includes is\nlisted, and only while it is there and ready.",
        "operationId": "list_shared_items",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "slides",
                    "audio",
                    "mindmap",
                    "notes"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`slides`, `audio`, `mindmap` or `notes`; every kind when absent",
              "title": "Kind"
            },
            "description": "`slides`, `audio`, `mindmap` or `notes`; every kind when absent"
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "description": "Case-insensitive; matched against the item's title, and its share as the feed matches it: the collection's title, the note and the cover's topic and terms",
              "default": "",
              "title": "Query"
            },
            "description": "Case-insensitive; matched against the item's title, and its share as the feed matches it: the collection's title, the note and the cover's topic and terms"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "newest",
                "reused"
              ],
              "type": "string",
              "description": "`newest` made first (the default) or `reused`: most reused share first",
              "default": "newest",
              "title": "Sort"
            },
            "description": "`newest` made first (the default) or `reused`: most reused share first"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 1,
              "default": 24,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedItems"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/{share_id}": {
      "get": {
        "tags": [
          "shares"
        ],
        "summary": "Get Share",
        "description": "One shared collection as a visitor sees it: what it is and what it\nincludes.",
        "operationId": "get_share",
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Share Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "shares"
        ],
        "summary": "Update Share",
        "description": "Change what your share includes; what is not sent stays as it is.",
        "operationId": "update_share",
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Share Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SharePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "shares"
        ],
        "summary": "Delete Share",
        "description": "Stop sharing a collection. Copies people already made stay theirs.",
        "operationId": "delete_share",
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Share Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/{share_id}/cover": {
      "get": {
        "tags": [
          "shares"
        ],
        "summary": "Read Share Cover",
        "description": "A shared collection's cover, as its owner sees it; see the collection's\ncover route.",
        "operationId": "read_share_cover",
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Share Id"
            }
          },
          {
            "name": "v",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "The `cover_version` the card was listed with",
              "default": "",
              "title": "V"
            },
            "description": "The `cover_version` the card was listed with"
          },
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "`light`, else the dark default",
              "default": "dark",
              "title": "Theme"
            },
            "description": "`light`, else the dark default"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/{share_id}/sources/{name}": {
      "get": {
        "tags": [
          "shares"
        ],
        "summary": "Read Shared Source",
        "description": "One source of a share that includes its sources, with its text.",
        "operationId": "read_shared_source",
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Share Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceText"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShareView": {
        "properties": {
          "card": {
            "$ref": "#/components/schemas/ShareCard"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceOut"
            },
            "type": "array",
            "title": "Sources",
            "description": "Empty when the sources are not included"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/SharedOutput"
            },
            "type": "array",
            "title": "Outputs",
            "description": "In the order the owner picked them"
          }
        },
        "type": "object",
        "required": [
          "card",
          "sources",
          "outputs"
        ],
        "title": "ShareView",
        "description": "One shared collection as a visitor sees it."
      },
...
      "SharedItems": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SharedItem"
            },
            "type": "array",
            "title": "Items"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset",
            "description": "Ask again with it as `offset` for the next page; null on the last"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_offset"
        ],
        "title": "SharedItems",
        "description": "One page of Discover's items."
      },
...
    }
  }
}